replace pycrypto with pyaes
Quick 'search and replace' to use pyaes which is a python native AES implementation.
this comes with a perfomance penalty but improves portability, e.g. homeassistant on windows
but more importantly for embedded (i.e. LEDE) where no pre-built modules are easily available and the platform isn't capable of building them. [and really because (I couldn't get) pycrypto to build using musl (iso glibc)]

I have tested this with an SP2 and an A1, I have no access to other devices for testing.

note: I don't speak 'python' (I've spent the best part of the morning hunting whitespace :( 
if it looks like kiddy-code it's because it is... feel free to improve.